requirements.txt
altair==5.2.0
jupyter==1.0.0
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
seaborn==0.13.2
streamlit==1.31.1
scikit-learn==1.4.1.post1
gdelt
google-cloud-bigquery
